major refactor but completely back compatible:  - removed all references to flair because it was becoming increasingly difficult to justify the separation between the servlet side of longwell and the rest (especially because piggy bank now needs to inject wrapped longwell service inside the servlet, which made flair and longwell coupled directly)  - abstracted the services provided by longwell to an interface, this will allow piggy bank to polymorphically pass its own overloaded version of services which are required to allow longwell to recognize the temporary profiles that piggy bank generates out of scraping)  - added methods to get and set the longwell profile on the longwell servlet (formerly flair)
There are still two references to 'flair' and mostly have to do with the facades (both RDF namespaces and file name on disk of the model that holds their information). Changing these will introduce back incompatibility and I didn't want to incur in such a penalty just for sake of naming coherence.