major refactoring in the way piggy bank works:
 1) now the data coin shows up on every web page, no matter what it contains... not only this is *a lot* faster (and avoids the creation of a thread and of a new triple store *PER PAGE LOADED* no matter if it was used or not, grrr), but also makes it more discoverable the fact that piggy bank is capable of scraping (and tagging!) pages that are just HTML.

 2) the javascript calls that take a considerable time are made non-blocking. This makes the browser feel a lot more responsive to user interaction (this also happens at startup). The downside is that piggybank won't be available for several seconds when the browser is starting up (this severely OS and JVM dependent too!) but again, makes the browser way more responsive during startup time.

In short, all this makes piggy bank a lot less intrusive in the user's browsing experience. Coupled with the new caching algorithm in longwell 2.4 this should really make piggy bank a reasonable extension to keep in your browser, even if you don't use it a lot (which is a necessary condition for people to learn to appreciate its presence).