You can now do a query and get results. (There are still plenty of debug print statements in the code.)
Integer IDs are turned to values one at a time; the result iterator
does a bunch of SQL queries every time you call next(). This is slow
and terrible. Really, we should be nesting every query in one of
the horrible all-at-once queries that getStatements() does. Implementing
this should make the rest of my week entertaining.