adding a triple store module for playground wrapping, therefore exposing, the vanilla Sesame2 API to Playground. This is useful because playground will use this API to store and retrieve its own internal information (like user info, their relationships, ACL and the like)
NOTE: one of the advantages of the sesame API is that the triple store can either be local or remote, but following the N>1 design pattern, I'm already expecting the triple store to be remote and accessed thru their HTTPRepository implementation. Since this makes it a little harder for other developers to start working on playground, I'll install a development version of sesame somewhere on our test machines soon so that that requirement will be lifted. The same will be true for the version control repository.